requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.119
botocore==1.34.119
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
jmespath==1.0.1
moviepy==1.0.3
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.3
pandas==2.2.2
pillow==10.3.0
proglog==0.1.10
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
tqdm==4.66.4
twilio==9.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
